Move bulk-user api to v5 as a separate resource
