+++ corehq/apps/api/resources/v0_5.py
@@ -1,9 +1,64 @@
+from collections import namedtuple
+
 from django.core.urlresolvers import reverse
+
+from tastypie import fields
 from tastypie.bundle import Bundle
-from corehq.apps.api.resources import v0_1, v0_4
+from tastypie.exceptions import BadRequest
+
 from corehq.apps.groups.models import Group
 from corehq.apps.sms.util import strip_plus
 from corehq.apps.users.models import CommCareUser, WebUser, CouchUser
+
+from ..es import UserESMixin
+from . import v0_1, v0_4
+from . import JsonResource, DomainSpecificResourceMixin
+
+
+class BulkUserResource(UserESMixin, JsonResource, DomainSpecificResourceMixin):
+    """
+    A read-only user data resource based on elasticsearch.
+    Supported Params: limit offset q fields filter?
+    Use cases: detail by id, detail by username, list by query
+    list syntax:
+        fields = ["first_name", "last_name"] == ?fields=first_name&fields=last_name
+    """
+    # UserESMixin containst the logic for interacting with ES
+    type = "bulk-user"
+    id = fields.CharField(attribute='id', readonly=True, unique=True)
+    email = fields.CharField(attribute='email')
+    username = fields.CharField(attribute='username', unique=True)
+    first_name = fields.CharField(attribute='first_name', null=True)
+    last_name = fields.CharField(attribute='last_name', null=True)
+    phone_numbers = fields.ListField(attribute='phone_numbers', null=True)
+
+    def to_obj(self, user):
+        '''
+        Takes a flat dict and returns an object
+        '''
+        if '_id' in user:
+            user['id'] = user.pop('_id')
+        return namedtuple('user', user.keys())(**user)
+
+    class Meta(v0_1.CustomResourceMeta):
+        list_allowed_methods = ['get']
+        detail_allowed_methods = ['get']
+        object_class = object
+        resource_name = 'bulk-user'
+
+    def obj_get_list(self, bundle, **kwargs):
+        # TODO: return only the specified fields
+        # fields = params.getlist('fields')
+
+        params = bundle.request.GET
+        param = lambda p: params.get(p, None)
+        users = self.make_query(
+                domain=kwargs['domain'],
+                q=param('q'),
+                size=param('limit'),
+                start_at=param('offset'),
+        )
+        return map(self.to_obj, users)
 
 
 class CommCareUserResource(v0_1.CommCareUserResource):
