+++ corehq/apps/api/urls.py
@@ -63,19 +63,7 @@ API_LIST = (
         v0_5.CommCareUserResource,
         v0_5.WebUserResource,
         v0_5.GroupResource,
-        FixtureResource,
-        ReportResource,
-    )),
-    ((0, 6), (
-        v0_4.ApplicationResource,
-        v0_4.CommCareCaseResource,
-        v0_4.XFormInstanceResource,
-        v0_4.RepeaterResource,
-        v0_4.SingleSignOnResource,
-        v0_4.HOPECaseResource,
-        v0_5.WebUserResource,
-        v0_5.GroupResource,
-        v0_6.CommCareUserResource,
+        v0_5.BulkUserResource,
         FixtureResource,
         ReportResource,
     )),
